==> jn setups/Step 5_1 Formalize Extensive Form Construction.ipynb <==
using Pkg
Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi

include("./parameters.jl")
include("./get_functions.jl")

loadcsv = CSV.File("./LOAD.csv");

include("./modification_functions.jl")

loaddis = load_distribution_dict(loadcsv);

# Below are test cases. Construction for getting and modifying uncertainty vectors (i.e. load and wind)
#     will be done elsewhere.

loadraw1 = [985.0197922,
985.7248887,
1001.58956,
1036.139287,
1139.788471,
1286.801089,
1347.086838,
1305.486145,
1232.508659,
1184.20955,
1153.185304,
1129.212024,
1107.706581,
1093.957199,
1087.611331,
1091.489362,
1170.460168,
1280.102672,
1266.000742,
1245.200396,
1204.657348,
1117.577932,
1031.203612,
991.0131123];

loadraw2 = [970.9178624,
968.4500247,
977.9688273,
1011.46091,
1097.83523,
1221.227115,
1274.109352,
1224.752598,
1174.690747,
1150.364918,
1128.506927,
1120.750866,
1112.642256,
1101.713261,
1099.245423,
1101.360713,
1163.056655,
1265.295646,
1256.129391,
1222.989857,
1176.10094,
1094.662296,
1009.345621,
963.8668976]

windraw1 = [713.2,
712.8,
708.4,
710.7,
701.4,
682.5,
614.7,
517.7,
426.6,
274.2,
93,
6.3,
3.8,
1.1,
0,
0,
0.9,
276.3,
272.9,
345.6,
411.7,
376.6,
561.3,
568.4];

windraw2 = [630,
672.3,
661.3,
654,
654.4,
666.4,
686.7,
689.3,
683.8,
698.6,
713.5,
710.3,
702.2,
699.2,
686.9,
695.4,
712.3,
712.4,
712.2,
711.2,
711.9,
696.7,
689.7,
666.6];

load = Dict()

load[1] = 0.01*loadraw1;
load[2] = 0.01*loadraw2;

wind = Dict()

wind[1] = 0.01*windraw1;
wind[2] = 0.01*windraw2;


nscen = 2

ptdfdict = ptdf_dict(ptdfdf);

ef, sub = construct_ef_model(load, wind, nscen)

JuMP.set_optimizer(ef, Gurobi.Optimizer)
JuMP.set_optimizer(sub, Gurobi.Optimizer)

optimize!(sub)

JuMP.optimize!(ef)

6.798147305e+01 - 6.798147305e+01
