==> jn setups/Step 7_3 look into LoL and OL.ipynb <==
using Pkg
Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames

include("./parameters.jl")
include("./get_functions.jl")
include("./modification_functions.jl")

model = JuMP.read_from_file("./ef_12scen_test_revised.mps")

JuMP.set_optimizer(model, Gurobi.Optimizer)
JuMP.set_optimizer_attribute(model, "OutputFlag", 1) 
JuMP.set_optimizer_attribute(model, "Method", 1) 

JuMP.optimize!(model)

for scen = 1:12
    for ts in timesteps
        for bus in buses
            var = get_lossofload_variable(model, bus, ts, scen)
            if value(var) > 0
                println(var, " ", value(var))
            end
        end
    end
end

for scen = 1:12
    for ts in timesteps
        for bus in buses
            var = get_overload_variable(model, bus, ts, scen)
            if value(var) > 0
                println(var, " ", value(var))
            end
        end
    end
end

for bus in buses
    varPR = get_PR_variable(model,bus,0)
    varER = get_ER_variable(model,bus,0)
    
    println(varPR, " ", value(varPR))
    println(varER, " ", value(varER))
end

0.15466402660380996+0.24943559054332543+4.566517218188891+5.379731685013425+2.925794031936479+0.09428337550546198+1.0069132191148722+2.9212059103401526+2.7909718971295505+3.0939258241139385+0.2978958550835288+1.3410492947362573

ans*100
