==> jn setups/Step 9 Compare Solutions.ipynb <==
using Pkg
Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames

include("./parameters.jl")
include("./get_functions.jl")
include("./modification_functions.jl")

model = JuMP.read_from_file("./ef_12scen_test_revised_noint.mps")

varnames = ["PR_{101}",
"ER_{101}",
"PR_{102}",
"ER_{102}",
"PR_{103}",
"ER_{103}",
"PR_{104}",
"ER_{104}",
"PR_{105}",
"ER_{105}",
"PR_{106}",
"ER_{106}",
"PR_{107}",
"ER_{107}",
"PR_{108}",
"ER_{108}",
"PR_{109}",
"ER_{109}",
"PR_{110}",
"ER_{110}",
"PR_{111}",
"ER_{111}",
"PR_{112}",
"ER_{112}",
"PR_{113}",
"ER_{113}",
"PR_{114}",
"ER_{114}",
"PR_{115}",
"ER_{115}",
"PR_{116}",
"ER_{116}",
"PR_{117}",
"ER_{117}",
"PR_{118}",
"ER_{118}",
"PR_{119}",
"ER_{119}",
"PR_{120}",
"ER_{120}",
"PR_{121}",
"ER_{121}",
"PR_{122}",
"ER_{122}",
"PR_{123}",
"ER_{123}",
"PR_{124}",
"ER_{124}"];


JuMP.set_optimizer(model, Gurobi.Optimizer)
JuMP.set_optimizer_attribute(model, "Method", 1) 

optimize!(model)

ef = 1.098147258e+03
multicut = 1.098605931e+03
vanilla = 1.098627357e+03
rd1sync = 1.098157008e+03

get_PR_lb(model,101,0)

soltest =  zeros(48)
#[0.01515317295046937, 0.07576586475234684, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.06992444968258622, 0.3496222484129311, 0.024425242996275243, 0.10597209391518082, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.09035704399269796, 0.4517852199634898, 0.0, 0.0, 0.0, 0.0, 0.01246785328231903, 0.06233926641159515, 0.0053185494034406755, 0.026592747017203378, 0.0, 0.0, 0.08532203616288067, 0.4266101808144034, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.2892514173319449, 1.4462570866597244, 0.0, 0.0]
for i = 1:48
    println(varnames[i], " ", soltest[i])
    JuMP.fix(variable_by_name(model, string(varnames[i],"_0")), soltest[i])
end

optimize!(model)

valtest = JuMP.objective_value(model)

(valtest-ef)/ef

valtest-ef

Theta = -26.287053667809914 
Q = 1571.1051347641887
Qt = 1592.6745769075285

Q - (Qt- 0.01*(Qt-Theta))

count = 1
for bus in buses
    varPR = get_PR_variable(model,bus,0)
    varER = get_ER_variable(model,bus,0)
    println(count, " ", varPR, " ", value(varPR))
    count += 1
    println(count, " ", varER, " ",value(varER))
    count += 1
end

regdec1 = [0.10148015190380144, 0.5074007595190072, 0.1054809953331612, 0.527404976665806, 0.0, 0.0, 0.11616734056181416, 0.5808367028090707, 0.1280583980698953, 0.6402919903494765, 0.14751071996050993, 0.7375535998025496, 0.0, 0.0, 0.14361017448783375, 0.7180508724391688, 0.12663440087148883, 0.6331720043574441, 0.043644225260060295, 0.21822112630030147, 0.06830918854124987, 0.34154594270624933, 0.061764138624737557, 0.30882069312368776, 0.07458710137299955, 0.3729355068649977, 0.12396169507410626, 0.6198084753705313, 0.0, 0.0, 0.21470391677366762, 1.073519583868338, 0.8219336633920721, 4.10966831696036, 0.3310062944375742, 1.6550314721878707, 0.18214283332098113, 0.9107141666049056, 0.14656286919129755, 0.7328143459564878, 0.12244751340207545, 0.6122375670103772, 0.2744545586844613, 1.3722727934223065, 0.13655682871955502, 0.6827841435977751, 0.0, 0.0];

for i = 1:48
    JuMP.fix(variable_by_name(model, string(varnames[i],"_0")), regdec1[i])
end

optimize!(model)

(1.098605931e+03 - 1.098147258e+03)/1.098147258e+03*100

vanillavals = [
    0.073171542,
0.365857712,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0.042605839,
0.213029195,
3.297463648,
16.48731824,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0.057775978,
0.288879892,
];

for i = 1:48
    JuMP.fix(variable_by_name(model, string(varnames[i],"_0")), vanillavals[i])
end

optimize!(model)

(rd1sync - ef)

regdecp01 = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.21503066912712618, 1.0751533456356308, 0.0, 0.0, 0.4436106743199207, 2.2180533715996034, 1.3589337263143402, 6.7946686315717, 0.0, 0.0, 0.24318918225253344, 1.2159459112626672, 0.529837882672972, 2.64918941336486, 0.0, 0.0, 0.07994478666409131, 0.3997239333204565, 0.6004700866325626, 3.002350433162813, 0.0, 0.0]
for i = 1:48
    JuMP.fix(variable_by_name(model, string(varnames[i],"_0")), regdecp01[i])
end

optimize!(model)

rdp01sync = 1.098149096e+03

(rdp01sync - ef)/ef

regdegp1 = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.039710180601863876, 0.19855090300931935, 0.0, 0.0, 0.0033450392283913724, 0.01672519614195686, 0.0, 0.0, 0.027311686066863097, 0.13655843033431547, 0.22420843444984884, 1.121042172249244, 0.0, 0.0, 0.5773692244701418, 2.886846122350709, 1.3108883122071608, 6.554441561035803, 0.1598497952849199, 0.7992489764245995, 0.4849745834297228, 2.424872917148614, 0.31181990636323864, 1.559099531816193, 0.0, 0.0, 0.08005897890473343, 0.4002948945236671, 0.25148086697643046, 1.2574043348821522, 0.0, 0.0]
for i = 1:48
    JuMP.fix(variable_by_name(model, string(varnames[i],"_0")), regdegp1[i])
end

optimize!(model)

rdp1sync = 1.098149221e+03

(rdp1sync - ef)/ef

regdeg10 = [0.10696488828794558, 0.5348244414397278, 0.10754571160481739, 0.5377285580240869, 0.08132683635859653, 0.4066341817929826, 0.1098848809027337, 0.5494244045136685, 0.11195222667175851, 0.5597611333587925, 0.11519486669038011, 0.5759743334519005, 0.0, 0.0, 0.1145266488256562, 0.572633244128281, 0.11156515810610644, 0.5578257905305322, 0.1187184707009062, 0.593592353504531, 0.12480986390643434, 0.6240493195321717, 0.12280509433112714, 0.6140254716556357, 0.12627552282061325, 0.6313776141030661, 0.1366466072282992, 0.6832330361414959, 0.06862354960543474, 0.34311774802717365, 0.15588725435087988, 0.7794362717543993, 0.4893553246621496, 2.4467766233107477, 0.29584269909065153, 1.4792134954532576, 0.14814442045910142, 0.7407221022955071, 0.1415414816903258, 0.7077074084516289, 0.20930336554840265, 1.0465168277420132, 0.27541501815353026, 1.3770750907676512, 0.13830407144251716, 0.6915203572125858, 0.060383046544971834, 0.30191523272485915]
for i = 1:48
    JuMP.fix(variable_by_name(model, string(varnames[i],"_0")), regdeg10[i])
end

optimize!(model)

rd10sync = 1.098165657e+03

(rd10sync - ef)/ef

regdegp1 = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.29158463267803847, 1.4579231633901923, 0.0, 0.0, 1.5448388211265778, 7.724194105632888, 1.469783051207154, 7.348915256035769, 0.0, 0.0, 0.08475152406696113, 0.4237576203348056, 0.0, 0.0, 0.0, 0.0, 0.08005897890460781, 0.40029489452304, 0.0, 0.0, 0.0, 0.0]
for i = 1:48
    JuMP.fix(variable_by_name(model, string(varnames[i],"_0")), regdegp1[i])
end

optimize!(model)

rdp001sync = 1.098147303e+03

(rdp001sync - ef)/ef

1098.1485584356633-ef

asynctest = [0,
0,
2.04E-10,
1.02E-09,
8.34E-06,
4.17E-05,
7.80E-06,
3.90E-05,
7.76E-06,
3.88E-05,
7.71E-06,
3.86E-05,
0,
0,
1.54E-05,
7.72E-05,
1.55E-05,
7.77E-05,
7.67E-06,
3.84E-05,
0,
0,
0,
0,
0.254447239,
1.272236196,
0.135226583,
0.676132916,
0,
0,
0.609701915,
3.048509575,
1.304256044,
6.521280221,
7.13E-05,
0.000356389,
0.266942611,
1.334713054,
0.434411529,
2.172057646,
0,
0,
0.07899352,
0.394967601,
0.386896021,
1.934480104,
0,
0];

for i = 1:48
    JuMP.fix(variable_by_name(model, string(varnames[i],"_0")), synctest[i])
end

optimize!(model)

(1.098149401e+03-ef)/ef

(1.098148557e+03-ef)/ef

synctest =[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.10645636619508117, 0.5322818309754058, 0.0, 0.0, 0.6390331051037744, 3.195165525518872, 1.474188427148458, 7.37094213574229, 0.0, 0.0, 0.30078165862765455, 1.5039082931382726, 0.426042511412889, 2.130212557064445, 0.0, 0.0, 0.0800589789046287, 0.40029489452314343, 0.4444559605909163, 2.2222798029545814, 0.0, 0.0]

multicutvals = [3.68E-12,
1.84E-11,
3.68E-12,
1.84E-11,
-3.68E-12,
-1.84E-11,
3.68E-12,
1.84E-11,
0,
0,
0.77701244,
3.885062201,
0,
0,
-3.69E-12,
-1.84E-11,
7.35E-12,
3.67E-11,
1.97272249,
9.863612451,
1.05E-09,
5.25E-09,
0.527091334,
2.635456671,
1.21E-08,
6.06E-08,
0,
0,
0,
0,
1.21E-08,
6.06E-08,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
0,
2.66E-11,
1.33E-10,
0.194190718,
0.97095359];

Qk = 1098.1551853765827 
Qtk = 1098.511430149813 
Tk = 1098.1343653482456

Qk = Qtk - gamma*(Qtk-Tk)

gamma = (Qk-Qtk)/(Tk-Qtk)
