==> jn setups/Step_10_get_worstcase_and_average_xvals.ipynb <==
using Pkg
Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames

include("./parameters.jl")
include("./get_functions.jl")
include("./modification_functions.jl")

budget = GenericAffExpr{Float64,VariableRef}()

varvals = Dict()

for i = 1:12
    varvals[i] = Dict()
    
    model = JuMP.read_from_file("./storage_expansion_revised/second_stage/noint_scen_$(i).mps")
    budget = GenericAffExpr{Float64,VariableRef}()
    for bus in buses
        varPR = get_PR_variable(model, bus)
        varER = get_ER_variable(model, bus)
        add_to_expression!(budget, 56.12, varPR)
        add_to_expression!(budget, 0.3, varER)
        @constraint(model, varPR >= 0.2*varER)
        @constraint(model, varPR <= 0.8*varER)
    end
    @constraint(model, budget <= 200.0)
    JuMP.set_optimizer(model, Gurobi.Optimizer)
    JuMP.set_optimizer_attribute(model, "Method", 1) 
    
    JuMP.optimize!(model)
    println(i, " ", objective_value(model))
    for bus in buses
        varPR = get_PR_variable(model, bus)
        varER = get_ER_variable(model, bus)
        varvals[i][name(varPR)] = value(varPR)
        varvals[i][name(varER)] = value(varER)
    end
    
end

# worst case
#7 8392.573477144413

varnames = ["PR_{101}",
"ER_{101}",
"PR_{102}",
"ER_{102}",
"PR_{103}",
"ER_{103}",
"PR_{104}",
"ER_{104}",
"PR_{105}",
"ER_{105}",
"PR_{106}",
"ER_{106}",
"PR_{107}",
"ER_{107}",
"PR_{108}",
"ER_{108}",
"PR_{109}",
"ER_{109}",
"PR_{110}",
"ER_{110}",
"PR_{111}",
"ER_{111}",
"PR_{112}",
"ER_{112}",
"PR_{113}",
"ER_{113}",
"PR_{114}",
"ER_{114}",
"PR_{115}",
"ER_{115}",
"PR_{116}",
"ER_{116}",
"PR_{117}",
"ER_{117}",
"PR_{118}",
"ER_{118}",
"PR_{119}",
"ER_{119}",
"PR_{120}",
"ER_{120}",
"PR_{121}",
"ER_{121}",
"PR_{122}",
"ER_{122}",
"PR_{123}",
"ER_{123}",
"PR_{124}",
"ER_{124}"];


vec7 = zeros(48)
    for j = 1:48
        vec7[j] += varvals[12][varnames[j]]
    end


show(vec7)

show(a0)

sumPR = 0.0
sumER = 0.0
for i = 1:48
    if i % 2 == 1
        println("1 ", i)
        sumPR += a0[i]
    else
        println("0 ", i)
        sumER += a0[i]
    end
end


sumPR

1 % 2

soltest = [0.0307608068430004, 0.22687940695123376, 0.11250476801672125, 0.9335139378571455, 0.14826470442029815, 1.1453174211018158, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0053185494034406755, 0.024517183112510655, 0.0, 0.0, 0.08532203616288067, 0.4674645993675957, 0.0, 0.0, 0.24369189074920142, 1.6908920586476954, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.16496582286915362, 1.1315926895534472, 0.024425242971849996, 0.10597209380920872, 0.0, 0.0]
for j = 1:12
    
    model = JuMP.read_from_file("./storage_expansion_revised/second_stage/noint_scen_$(j).mps")
    JuMP.set_optimizer(model, Gurobi.Optimizer)
    JuMP.set_optimizer_attribute(model, "Method", 1) 
        
    for i = 1:48
        JuMP.fix(variable_by_name(model, string(varnames[i])), soltest[i])
    end
    
    JuMP.optimize!(model)
end
