==> jn setups/Step 11 Scenario IDs.ipynb <==
using Random

Random.seed!(5)

wholevec = shuffle(1:6940)

validation = wholevec[1:4000]

trainingvec = wholevec[4001:6940]

strat1458 = trainingvec[1:1458]
strat723 = trainingvec[1:723]
strat281 = trainingvec[1:281]
strat133 = trainingvec[1:133]

shuffidx1458 = shuffle(1:2940)[1:1458]
shuffidx723 = shuffle(1:2940)[1:723]
shuffidx281 = shuffle(1:2940)[1:281]
shuffidx133 = shuffle(1:2940)[1:133]

shuff1458 = Int64[]

for i = 1:1458
    push!(shuff1458, trainingvec[shuffidx1458[i]])
end

shuff723 = Int64[]

for i = 1:723
    push!(shuff723, trainingvec[shuffidx723[i]])
end

shuff281 = Int64[]

for i = 1:281
    push!(shuff281, trainingvec[shuffidx281[i]])
end

shuff133 = Int64[]

for i = 1:133
    push!(shuff133, trainingvec[shuffidx133[i]])
end

part1458 = trainingvec[1:1458]
part723 = trainingvec[1459:2181]
part281 = trainingvec[2182:2462]
part133 = trainingvec[2463:2595]

println("wholevec = ", show(wholevec))

show(validation)

show(strat1458)

show(strat723)

show(strat281)

show(strat133)

show(shuff1458)

show(shuff723)

show(shuff281)

show(shuff133)

show(part1458)

show(part723)

show(part281)

show(part133)
