==> jn setups/Step_12_Get_Hourly_Data_Into_One_Source.ipynb <==
using DataFrames
using CSV

# Load CSV

df2002 = DataFrame(CSV.File("../FPdatascratch\\2002_ercot_hourly_load_data.csv"))
df2003 = DataFrame(CSV.File("../FPdatascratch\\2003_ercot_hourly_load_data.csv"))
df2004 = DataFrame(CSV.File("../FPdatascratch\\2004_ercot_hourly_load_data.csv"))
df2005 = DataFrame(CSV.File("../FPdatascratch\\2005_ercot_hourly_load_data.csv"))
df2006 = DataFrame(CSV.File("../FPdatascratch\\2006_ercot_hourly_load_data.csv"))
df2007 = DataFrame(CSV.File("../FPdatascratch\\2007_ercot_hourly_load_data.csv"))
df2008 = DataFrame(CSV.File("../FPdatascratch\\2008_ercot_hourly_load_data.csv"))
df2009 = DataFrame(CSV.File("../FPdatascratch\\2009_ercot_hourly_load_data.csv"))
df2010 = DataFrame(CSV.File("../FPdatascratch\\2010_ercot_hourly_load_data.csv"))
df2011 = DataFrame(CSV.File("../FPdatascratch\\2011_ercot_hourly_load_data.csv"))
df2012 = DataFrame(CSV.File("../FPdatascratch\\2012_ercot_hourly_load_data.csv"))
df2013 = DataFrame(CSV.File("../FPdatascratch\\2013_ercot_hourly_load_data.csv"))
df2014 = DataFrame(CSV.File("../FPdatascratch\\2014_ercot_hourly_load_data.csv"))
df2015 = DataFrame(CSV.File("../FPdatascratch\\native_load_2015.csv"))
df2016 = DataFrame(CSV.File("../FPdatascratch\\native_Load_2016.csv"))
df2017 = DataFrame(CSV.File("../FPdatascratch\\native_Load_2017.csv"))
df2018 = DataFrame(CSV.File("../FPdatascratch\\Native_Load_2018.csv"))
df2019 = DataFrame(CSV.File("../FPdatascratch\\Native_Load_2019.csv"))
df2020 = DataFrame(CSV.File("../FPdatascratch\\Native_Load_2020.csv"))

size(df2016)

dfload = DataFrame(scenid = Int64[], date= String[], 
    h1 = Float64[], 
    h2 = Float64[], 
    h3 = Float64[],
    h4 = Float64[],
    h5 = Float64[],
    h6 = Float64[],
    h7 = Float64[],
    h8 = Float64[],
    h9 = Float64[],
    h10 = Float64[],
    h11 = Float64[],
    h12 = Float64[],
    h13 = Float64[], 
    h14 = Float64[], 
    h15 = Float64[],
    h16 = Float64[],
    h17 = Float64[],
    h18 = Float64[],
    h19 = Float64[],
    h20 = Float64[],
    h21 = Float64[],
    h22 = Float64[],
    h23 = Float64[],
    h24 = Float64[])

#2002
count = 0
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2002[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2002[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2003
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2003[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2003[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2004
for i = 0:365
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2004[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2004[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2005
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2005[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2005[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2006
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2006[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2006[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2007
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2007[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2007[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2008
for i = 0:365
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2008[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2008[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2009
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2009[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2009[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2010
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2010[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2010[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2011
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2011[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2011[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2012
for i = 0:365
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2012[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2012[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2013
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2013[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2013[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2014
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2014[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2014[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2015
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2015[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2015[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2016
for i = 0:365
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2016[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2016[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2017
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2017[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2017[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2018
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2018[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2018[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2019
for i = 0:364
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2019[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2019[24*i+j,11])
    end
    push!(dfload, scenvec)
end

#2020
for i = 0:365
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, df2020[24*i+1,1])
    for j = 1:24
        push!(scenvec, df2020[24*i+j,11])
    end
    push!(dfload, scenvec)
end

dfload

for i = 1:6939
    if dfload[i,2] == dfload[i+1,2]
        println("uh oh!")
    end
end

for i = 1:6940
    if dfload[i,1] != i
        println("uh oh!")
    end
end

CSV.write("../FPdatascratch\\loaddata.csv", dfload)

# Wind CSV

dfwold = DataFrame(CSV.File("../FPdatascratch\\single_wind_long_form.csv"))

dfwnew = DataFrame(scenid = Int64[], date= String[], 
    h1 = Float64[], 
    h2 = Float64[], 
    h3 = Float64[],
    h4 = Float64[],
    h5 = Float64[],
    h6 = Float64[],
    h7 = Float64[],
    h8 = Float64[],
    h9 = Float64[],
    h10 = Float64[],
    h11 = Float64[],
    h12 = Float64[],
    h13 = Float64[], 
    h14 = Float64[], 
    h15 = Float64[],
    h16 = Float64[],
    h17 = Float64[],
    h18 = Float64[],
    h19 = Float64[],
    h20 = Float64[],
    h21 = Float64[],
    h22 = Float64[],
    h23 = Float64[],
    h24 = Float64[])

#2002
count = 0
for i = 0:6939
    count += 1
    scenvec = [];
    push!(scenvec, count)
    push!(scenvec, dfwold[24*i+1,1])
    for j = 1:24
        push!(scenvec, dfwold[24*i+j,3])
    end
    push!(dfwnew, scenvec)
end

CSV.write("../FPdatascratch\\winddata.csv", dfwnew)
