==> jn setups/Step_13_1 Finetune models decomposed.ipynb <==
using Pkg
Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames
using Random

include("./parameters.jl")
include("./get_functions.jl")
include("./modification_functions.jl")

# Experiment 0: load in first stage model

model = JuMP.read_from_file("./storage_expansion_revised/first_stage/first_stage_model_PR.mps")

# Experiment 1: expand budget to 1 Billion

exbud = get_expansion_budget(model)

set_normalized_rhs(exbud, 1000.0)

exbud

JuMP.write_to_file(model, "./storage_expansion_revised/first_stage/first_stage_model_exp1.mps")

count = 1
PRsum = 0.0
ERsum = 0.0
for bus in buses
    varPR = get_PR_variable(efmodel,bus,0)
    varER = get_ER_variable(efmodel,bus,0)
    println(count, " ", varPR, " ", value(varPR))
    PRsum += value(varPR)
    count += 1
    println(count, " ", varER, " ",value(varER))
    ERsum += value(varER)
    count += 1
end

ERsum

PRsum

# Experiment 2: adjust FS cost by up to 1.15%

Random.seed!(6)

busorder = shuffle(101:124)

adjustment = Float64[];
for i = 1:24
    push!(adjustment, (-0.0115+(i-1)*0.001)/100)
    println(adjustment[i])
end

busadjdict = Dict()

for i = 1:24
    println(busorder[i], " ", adjustment[i])
    busadjdict[busorder[i]] = adjustment[i]
end

objfunc = objective_function(model);
exbud = get_expansion_budget(model);

for bus in buses
    varPR = get_PR_variable(model, bus)
    #varER = get_ER_variable(model, bus)
    oldPR = coefficient(objfunc, varPR)
    #oldER = coefficient(objfunc, varER)
    
    set_objective_coefficient(model, varPR, (1+busadjdict[bus])*oldPR)
    #set_objective_coefficient(model, varER, (1+busadjdict[bus])*oldER)
    
    set_normalized_coefficient(exbud, varPR, (1+busadjdict[bus])*oldPR)
    #set_normalized_coefficient(exbud, varER, (1+busadjdict[bus])*oldER)
    
end
    

exbud

JuMP.write_to_file(model, "./storage_expansion_revised/first_stage/first_stage_model_PR_exp2p00015.mps")

# Experiment 3: The hourly ramp rates of generators are set to 45% of the maximum generation for the largest units in the system. Yan et al.

I define largest as > 100 MW max power

gencap = Dict()

caps = [20.,20,76,76,20,20,76,76,355,55,55,55,55,12,12,155,155,355,155,350,55,55,55,400]

for i = 1:24
    gencap[gens[i]] = caps[i]
end

for (key, value) in gencap
    println(key, " ", value)
end

for scen in 1:12
    model = JuMP.read_from_file("./storage_expansion_revised/second_stage/noint_PR_scen_$(scen).mps") 
    for gen in gens
        if gencap[gen] > 100.0
            for ts in timesteps                
                if ts > 1
                    lbrampcon = get_ramp_lb(model, gen, ts)
                    ubrampcon = get_ramp_ub(model, gen, ts)
                    
                    set_normalized_rhs(lbrampcon, -0.45*gencap[gen]/100)
                    set_normalized_rhs(ubrampcon, 0.45*gencap[gen]/100)
                end
            end
        end
    end
    JuMP.write_to_file(model, "./storage_expansion_revised/second_stage/noint_PR_exp3_scen_$(scen).mps") 
end

# Experiment 4 (to replace 2): decrease branch capacity to 50% of original

for sc in 1:12
    model = JuMP.read_from_file("./storage_expansion_revised/second_stage/noint_scen_$(sc).mps") 
    for br in branches
        for ts in timesteps
            conub = get_branch_ub(model, br, ts)
            conlb = get_branch_lb(model, br, ts)
            
            oldub = JuMP.constraint_object(conub).set.upper
            oldlb = JuMP.constraint_object(conlb).set.lower
            JuMP.set_normalized_rhs(conub, 5/8*oldub)
            JuMP.set_normalized_rhs(conlb, 5/8*oldlb)
            
        end
    end
    JuMP.write_to_file(model, "./storage_expansion_revised/second_stage/noint_scen_exp4_$(sc).mps") 
end

# Experiment 34

for sc in 1:12
    model = JuMP.read_from_file("./storage_expansion_revised/second_stage/noint_scen_$(sc).mps") 
    #exp3
    for gen in gens
        if gencap[gen] > 100.0
            for ts in timesteps                
                if ts > 1
                    lbrampcon = get_ramp_lb(model, gen, ts)
                    ubrampcon = get_ramp_ub(model, gen, ts)
                    
                    set_normalized_rhs(lbrampcon, -0.45*gencap[gen]/100)
                    set_normalized_rhs(ubrampcon, 0.45*gencap[gen]/100)
                end
            end
        end
    end
    
    #exp 4
    for br in branches
        for ts in timesteps
            conub = get_branch_ub(model, br, ts)
            conlb = get_branch_lb(model, br, ts)
            
            oldub = JuMP.constraint_object(conub).set.upper
            oldlb = JuMP.constraint_object(conlb).set.lower
            JuMP.set_normalized_rhs(conub, 5/8*oldub)
            JuMP.set_normalized_rhs(conlb, 5/8*oldlb)
            
        end
    end
    JuMP.write_to_file(model, "./storage_expansion_revised/second_stage/noint_scen_exp34_$(sc).mps") 
end
