==> jn setups/Step_14 Remove ER variables.ipynb <==
using Pkg
Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames
using Random

include("./parameters.jl")
include("./get_functions.jl")
include("./modification_functions.jl")

model = JuMP.read_from_file("./storage_expansion_revised/first_stage/first_stage_model.mps")

model2 = [];
for i = 1:12
    push!(model2, JuMP.read_from_file("./storage_expansion_revised/second_stage/noint_scen_$(i).mps") )
end

for bus in buses
    JuMP.delete(model,get_ER_variable(model,bus))
end

for bus in buses
    JuMP.set_objective_coefficient(model, get_PR_variable(model, bus), 57.62)
    JuMP.delete(model, get_PR_ub(model,bus))
    JuMP.delete(model, get_ER_lb(model,bus))
end

budget = get_expansion_budget(model)

for bus in buses
    JuMP.set_coefficient(budget, get_PR_variable(model, bus), 57.62)
end

budget

# For Exp 1 only

set_normalized_rhs(budget, 1000.0)

budget

# end

for i = 1:12
    for bus in buses
        JuMP.delete(model2[i],get_ER_variable(model2[i],bus))
        JuMP.set_objective_coefficient(model2[i], get_PR_variable(model2[i], bus), 57.62)
        
        var = get_PR_variable(model2[i], bus)
        for ts in timesteps
            con = get_storage_ub(model2[i],bus,ts)
            set_coefficient(con, var, -5.0)
        end
    end
end

# for experiment 4 only

for sc in 1:12
    for br in branches
        for ts in timesteps
            conub = get_branch_ub(model2[sc], br, ts)
            conlb = get_branch_lb(model2[sc], br, ts)
            
            oldub = JuMP.constraint_object(conub).set.upper
            oldlb = JuMP.constraint_object(conlb).set.lower
            JuMP.set_normalized_rhs(conub, 5/8*oldub)
            JuMP.set_normalized_rhs(conlb, 5/8*oldlb)
            
        end
    end
end

JuMP.write_to_file(model, "./storage_expansion_revised/first_stage/first_stage_model_PR_exp1.mps")

for i = 1:12
    JuMP.write_to_file(model2[i], "./storage_expansion_revised/second_stage/noint_PR_exp4_scen_$(i).mps")
end

Random.seed!(6)
busorder = shuffle(101:124)

adjustment = Float64[];
for i = 1:24
    push!(adjustment, (-0.0115+(i-1)*0.001)/100)
    println(adjustment[i])
end

busadjdict = Dict()

for i = 1:24
    println(busorder[i], " ", adjustment[i])
    busadjdict[busorder[i]] = adjustment[i]
end

objfunc = objective_function(model);
exbud = get_expansion_budget(model);

for bus in buses
    varPR = get_PR_variable(model, bus)
    #varER = get_ER_variable(model, bus)
    oldPR = coefficient(objfunc, varPR)
    #oldER = coefficient(objfunc, varER)
    
    set_objective_coefficient(model, varPR, (1+busadjdict[bus])*oldPR)
    #set_objective_coefficient(model, varER, (1+busadjdict[bus])*oldER)
    
    set_normalized_coefficient(exbud, varPR, (1+busadjdict[bus])*oldPR)
    #set_normalized_coefficient(exbud, varER, (1+busadjdict[bus])*oldER)
    
end
    

# extensive_form

efmodel = JuMP.read_from_file("ef_12scen_test_revised_noint.mps")

JuMP.set_optimizer(efmodel, Gurobi.Optimizer)
JuMP.set_optimizer_attribute(efmodel, "Method", 1) 

for bus in buses
    JuMP.delete(efmodel,get_ER_variable(efmodel,bus,0))
end

for bus in buses
    JuMP.set_objective_coefficient(efmodel, get_PR_variable(efmodel, bus,0), 57.62)
    JuMP.delete(efmodel, get_PR_ub(efmodel,bus,0))
    JuMP.delete(efmodel, get_ER_lb(efmodel,bus,0))
end

budget = get_expansion_budget(efmodel,0)

for bus in buses
    JuMP.set_coefficient(budget, get_PR_variable(efmodel, bus,0), 57.62)
end

budget

for i = 1:12
    for bus in buses
        #JuMP.delete(efmodel,get_ER_variable(efmodel,bus,i))
        #JuMP.set_objective_coefficient(efmodel, get_PR_variable(efmodel, bus,0), 57.62)
        
        var = get_PR_variable(efmodel, bus,0)
        for ts in timesteps
            con = get_storage_ub(efmodel,bus,ts,i)
            set_coefficient(con, var, -5.0)
        end
    end
end

optimize!(efmodel)

write_to_file(efmodel, "ef_12scen_test_revised_noint_PR.mps")

Random.seed!(6)
busorder = shuffle(101:124)

adjustment = Float64[];
for i = 1:24
    push!(adjustment, (-0.0115+(i-1)*0.001)/100)
    println(adjustment[i])
end

busadjdict = Dict()

for i = 1:24
    println(busorder[i], " ", adjustment[i])
    busadjdict[busorder[i]] = adjustment[i]
end

objfunc = objective_function(efmodel);
exbud = get_expansion_budget(efmodel,0);

for bus in buses
    varPR = get_PR_variable(efmodel, bus, 0)
    #varER = get_ER_variable(efmodel, bus, 0)
    oldPR = coefficient(objfunc, varPR)
    #oldER = coefficient(objfunc, varER)
    
    set_objective_coefficient(efmodel, varPR, (1+busadjdict[bus])*oldPR)
    #set_objective_coefficient(efmodel, varER, (1+busadjdict[bus])*oldER)
    
    set_normalized_coefficient(exbud, varPR, (1+busadjdict[bus])*oldPR)
    #set_normalized_coefficient(exbud, varER, (1+busadjdict[bus])*oldER)
    
end
    

exbud

optimize!(efmodel)
