==> jn setups/Step 15 add maximal budget (save for in model adjustments).ipynb <==
using Pkg
Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames
using Random

include("./parameters.jl")
include("./get_functions.jl")
include("./modification_functions.jl")

loaddis = load_distribution_dict(loadcsv);

ptdfdf = DataFrame(CSV.File("./ptdfsmall.csv"));
ptdfdict = Dict()

for i = 1:38
    br = ptdfdf[i,1]
    ptdfdict[br] = Dict()
    for j = 2:25
        bus = parse(Int64,names(ptdfdf)[j])
        ptdfdict[br][bus] = ptdfdf[i,j]
    end
end

loadraw = 1.35*2850*ones(24);
for i = 1:3
   loadraw[i] = 0
end
windraw = zeros(24);


load = 0.01*loadraw
wind = 0.01*windraw
load

submodel = JuMP.read_from_file("./storage_expansion_revised/second_stage/noint_PR_scen_1.mps")

JuMP.set_optimizer(submodel, Gurobi.Optimizer)
JuMP.set_optimizer_attribute(submodel, "Method", 1) 

#adjust load balance values (remember to change load)

for bus in buses
    lf = loaddis[bus]
    for ts in timesteps
        con = get_load_balance(submodel, bus, ts)
        oldval = JuMP.constraint_object(con).set.value
        lval = load[ts]
        JuMP.set_normalized_rhs(con, lf*lval)
        newval = JuMP.constraint_object(con).set.value
        println("$(name(con)), $(oldval), $(newval)")
    end
end

# change ptdf constraint (remember to run load changes FIRST)
for ts in timesteps
    for br in branches
        ptdfcon = get_ptdf_con(submodel,br,ts)
        
        valold = JuMP.constraint_object(ptdfcon).set.value
        valnew = 0.0
        for bus in buses
            buscon = get_load_balance(submodel,bus,ts)
            
            loadcon = copy(JuMP.constraint_object(buscon).func)
            loadval = copy(JuMP.constraint_object(buscon).set.value)

            valnew -= ptdfdict[br][bus]*loadval

        end 

        JuMP.set_normalized_rhs(ptdfcon, valnew)
        println("$(JuMP.name(ptdfcon)), $(valold), $(valnew)")
        
    end
end


#adjust wind ub values (remember to change wind)

bus = 122
for ts in timesteps
    con = get_wind_ub(submodel, bus, ts)
    oldval = JuMP.constraint_object(con).set.upper
    wval = wind[ts]
    JuMP.set_normalized_rhs(con, wval)
    newval = JuMP.constraint_object(con).set.upper
    println("$(name(con)), $(oldval), $(newval)")
end


for bus in buses
    unfix(get_stored_variable(submodel,bus,1))
    unfix(get_stored_variable(submodel,bus,24))
    delete(submodel,get_storage_balance(submodel,bus,1))
end

optimize!(submodel)

sum = 0.
for i=101:124
    var = get_PR_variable(submodel,i)
    println(var, " ", value(var))
    sum+= value(var)
end
57.62*sum

is_fixed(get_stored_variable(submodel,101,24))

model1 = JuMP.read_from_file("./storage_expansion_revised/first_stage/first_stage_model_PR.mps")

budget = get_expansion_budget(model1)

set_normalized_rhs(budget, 3500)

budget

JuMP.write_to_file(model1, "./storage_expansion_revised/first_stage/first_stage_model_PR_exp1B.mps")
