==> jn setups/Step 16 More Degeneracy attempts.ipynb <==
using Pkg
Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames
using Random

include("./parameters.jl")
include("./get_functions.jl")
include("./modification_functions.jl")

Random.seed!(6)
busorder = shuffle(101:124)

adjustment = Float64[];
for i = 1:24
    push!(adjustment, (-0.0115+(i-1)*0.001)/100)
    println(adjustment[i])
end

busadjdict = Dict()

for i = 1:24
    println(busorder[i], " ", adjustment[i])
    busadjdict[busorder[i]] = adjustment[i]
end


efmodel = read_from_file("ef_12scen_test_revised_noint_PR.mps")
JuMP.set_optimizer(efmodel, Gurobi.Optimizer)
JuMP.set_optimizer_attribute(efmodel, "Method", 1) 

model2 = [];
for i = 1:1#2
    push!(model2, JuMP.read_from_file("./storage_expansion_revised/second_stage/noint_PR_exp3_scen_$(i).mps") )
end

budget = get_expansion_budget(efmodel, 0)

set_normalized_rhs(budget, 3500)

budget

#=for sc = 1:12
    for bus in buses
        for ts in timesteps
            varch = get_charging_variable(efmodel, bus, ts, sc)
            vardi = get_discharging_variable(efmodel, bus, ts, sc)
            
            con = get_storage_balance(efmodel, bus, ts, sc)
            
            coeffch = normalized_coefficient(con, varch)
            coeffdi = normalized_coefficient(con, vardi)
            
            set_normalized_coefficient(con, varch, (1+busadjdict[bus])*coeffch)
            set_normalized_coefficient(con, vardi, coeffdi/(1+busadjdict[bus]))
            
        end
    end
end=#

get_charge_ub(efmodel, 101, 1,1)

get_discharge_ub(efmodel, 101, 1, 1)

get_storage_ub(efmodel, 101, 1, 1)

for sc = 1:12
    for bus in buses
        var = get_PR_variable(efmodel, bus, 0)
        for ts in timesteps
            conc = get_charge_ub(efmodel, bus, ts, sc)
            cond = get_discharge_ub(efmodel, bus, ts, sc)
            cone = get_storage_ub(efmodel, bus, ts, sc)
            
            valc = JuMP.normalized_coefficient(conc, var)
            vald = JuMP.normalized_coefficient(cond, var)
            vale = JuMP.normalized_coefficient(cone, var)
            
            JuMP.set_normalized_coefficient(conc, var, (1+busadjdict[bus])*valc)
            JuMP.set_normalized_coefficient(cond, var, (1+busadjdict[bus])*vald)
            JuMP.set_normalized_coefficient(cone, var, (1+busadjdict[bus])*vale)
        end
    end
end

optimize!(efmodel)

sum = 0.
for bus in buses
    var = get_PR_variable(efmodel, bus, 0)
    println(var, " ",value(var))
    sum+=value(var)
end
println(sum)
println()
println(objective_value(efmodel))

#=for sc = 1:12
    for bus in buses
        for ts in timesteps
            varch = get_charging_variable(model2[sc], bus, ts)
            vardi = get_discharging_variable(model2[sc], bus, ts)
            
            con = get_storage_balance(model2[sc], bus, ts)
            
            coeffch = normalized_coefficient(con, varch)
            coeffdi = normalized_coefficient(con, vardi)
            
            set_normalized_coefficient(con, varch, (1+busadjdict[bus])*coeffch)
            set_normalized_coefficient(con, vardi, coeffdi/(1+busadjdict[bus]))
            
        end
    end
end=#

for sc = 1:1#2
    for bus in buses
        var = get_PR_variable(model2[sc], bus)
        for ts in timesteps
            conc = get_charge_ub(model2[sc], bus, ts)
            cond = get_discharge_ub(model2[sc], bus, ts)
            cone = get_storage_ub(model2[sc], bus, ts)
            
            valc = JuMP.normalized_coefficient(conc, var)
            vald = JuMP.normalized_coefficient(cond, var)
            vale = JuMP.normalized_coefficient(cone, var)
            
            JuMP.set_normalized_coefficient(conc, var, (1+busadjdict[bus])*valc)
            JuMP.set_normalized_coefficient(cond, var, (1+busadjdict[bus])*vald)
            JuMP.set_normalized_coefficient(cone, var, (1+busadjdict[bus])*vale)
        end
    end
end

#get_storage_balance(model2[7], 101, 1)

println(get_charge_ub(model2[1], 101, 1))
println(get_discharge_ub(model2[1], 101, 1))
println(get_storage_ub(model2[1], 101, 1))

for i = 1:1#2
    JuMP.write_to_file(model2[i], "./storage_expansion_revised/second_stage/noint_PR_exp36p00015_scen_$(i).mps") 
end

for bus in buses
    println(bus, " ", busadjdict[bus])
end

for (key, value) in objective_function(model2[7]).terms
    println(key, " ", value)
end
