==> jn setups/Step 19 Get Info from Validation.ipynb <==
using DataFrames
using CSV

dir = "../four_wind_turbine_experiments/validation/benchmarks_comp/benchmarks/"
traininfo = DataFrame(CSV.File(string(dir,"training_results_summary.csv")))

validinfo = string(dir, "validation_summary.csv")

#arrayid = 1
for arrayid = 1:size(traininfo,1)
    variation = traininfo[arrayid, 2]
    budget = traininfo[arrayid, 3]
    exp = traininfo[arrayid, 4]
    nscens = traininfo[arrayid, 5]
    mod = traininfo[arrayid, 6]
    objvaltr = traininfo[arrayid,7]
    fsval = 57.62*sum(traininfo[arrayid,8:31])
    
    println("$(arrayid), $(variation), $(budget), $(exp), $(nscens), $(mod), $(objvaltr)")
    
    solsumfile = string(dir, "v$(variation)_b$(budget)_e$(exp)_n$(nscens)_$(mod)/solution_summary.csv")
    
    solsumdf = DataFrame(CSV.File(solsumfile)) 
    
    vec = Vector{Any}(undef, 133)
    vec[1] = traininfo[arrayid,1]
    vec[2] = variation
    vec[3] = budget
    vec[4] = exp
    vec[5] = nscens
    vec[6] = mod
    vec[7] = objvaltr
    vec[8] = fsval
    vec[9] = sum(solsumdf[:,2])/4000
    
    vec[10] = maximum(solsumdf[:,2])
    
    tcv = 0
    la = 0
    oa = 0
    for i = 1:4000
        if solsumdf[i,2] > objvaltr
            tcv += 1
        end
        if sum(solsumdf[i,76:99])>0.0
            la += 1
        end
        if sum(solsumdf[i,100:123])>0.0
            oa += 1
        end
    end
    
    vec[11] = tcv
    vec[12] = la
    vec[13] = oa
    
    for i = 14:133
        vec[i] = sum(solsumdf[:,i-10])
    end
    
    #for i = 1:124
    #    println(i, " ", vec[i])
    #end
    
    vecstring = string(vec)
    n=length(vecstring)
    vecstring = vecstring[2:n-1]
    open(validinfo, "a") do io
        write(io, "$(vecstring) \n")
    end
    
end
            
    
    
    
    #mkdir(dir)
    #cp("./validation_results/solution_summary.csv", string(dir, "solution_summary.csv"))
#end

#arrayid = 1
for arrayid = 1:size(solinfo,1)
    variation = solinfo[arrayid, 2]
    budget = solinfo[arrayid, 3]
    exp = solinfo[arrayid, 4]
    nscens = solinfo[arrayid, 5]
    mod = solinfo[arrayid, 6]
    objvaltr = solinfo[arrayid,7]
    
    #println("$(arrayid), $(variation), $(budget), $(exp), $(nscens), $(mod), $(objvaltr)")
    
    solsumfile = "../results_and_validation/validation_results_so_1en5/validation_results_so_1en5/validation_results_so_1en5/v$(variation)_b$(budget)_e$(exp)_n$(nscens)_$(mod)/solution_summary.csv"
    

    
    solsumdf = DataFrame(CSV.File(solsumfile)) 
    
    summ = 0.0
    max = 0.0
    mind = 0
    for i = 1:4000
        summ += sum(solsumdf[i,76:99])
        if sum(solsumdf[i,76:99]) > max
            max = sum(solsumdf[i,76:99])
            mind = i
        end
    end
    avg = summ/4000
    println(arrayid, ", ", avg, ", ", max)#, ", ", mind, ", ", findmax(solsumdf[:,2])[2])
    #=
    vec = Vector{Float64}(undef, 125)
    
    vec[1] = sum(solsumdf[:,2])/4000
    
    vec[2] = maximum(solsumdf[:,2])
    
    tcv = 0
    la = 0
    oa = 0
    for i = 1:4000
        if solsumdf[i,2] > objvaltr
            tcv += 1
        end
        if sum(solsumdf[i,76:99])>0.0
            la += 1
        end
        if sum(solsumdf[i,100:123])>0.0
            oa += 1
        end
    end
    
    vec[3] = tcv
    vec[4] = la
    vec[5] = oa
    
    for i = 6:125
        vec[i] = sum(solsumdf[:,i-2])
    end
    
    #for i = 1:124
    #    println(i, " ", vec[i])
    #end
    
    vecstring = string(vec)
    n=length(vecstring)
    vecstring = vecstring[2:n-1]
    open(valsum, "a") do io
        write(io, "$(vecstring) \n")
    end
    =#
end
            
    
    
    
    #mkdir(dir)
    #cp("./validation_results/solution_summary.csv", string(dir, "solution_summary.csv"))
#end
