==> jn setups/Step 20 Build Single Scenario EF.ipynb <==
#using Pkg
#Pkg.activate(".")

using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames
using Random
using LinearAlgebra

include("../get_functions.jl")




include("../LShaped.jl/src/structs.jl")
include("../LShaped.jl/src/setup.jl")
include("../LShaped.jl/src/algorithm.jl")

include("./parameters.jl")
include("./modification_functions_async.jl")
loadcsv = CSV.File("./LOAD.csv");
ptdfdf = DataFrame(CSV.File("./ptdfsmall.csv"));
loaddf = DataFrame(CSV.File("./loaddata.csv"))
winddf = DataFrame(CSV.File("./winddata.csv"))

ercotscens133 = collect(DataFrame(CSV.File("./scenarios/part133.csv"))[1,:])

ercotscens = ercotscens133[1:12]

loaddf

let 
    lmax = 0.0
    wmax = 0.0
    for i = 1:24
        #println(i, " ", maximum(loaddf[:,2+i]))
        if maximum(loaddf[:,2+i]) > lmax
            lmax = maximum(loaddf[:,2+i])
        end
        #println(maximum(winddf[:,2+i]))
        if maximum(winddf[:,2+i]) > wmax
            wmax = maximum(winddf[:,2+i])
        end
    end
    global lmax = lmax
    global wmax = wmax
end

loaddis = load_distribution_dict(loadcsv);

ptdfdict = Dict()

for i = 1:38
    br = ptdfdf[i,1]
    ptdfdict[br] = Dict()
    for j = 2:25
        bus = parse(Int64,names(ptdfdf)[j])
        ptdfdict[br][bus] = ptdfdf[i,j]
    end
end

lrts = 2850.0
wrts = 713.5

model2nd = JuMP.read_from_file("../storage_expansion_revised/ercot/PR_exp3_scen_1.mps")

var = get_lossofload_variable(model2nd, 101, 1)

objval = objective_function(model2nd);

#var = get_discharging_variable(model2nd, 101,1)

coeff = objval.terms[var]

coeff*1000000/100/365.25

avg*fuel_cost/100000

vardict =  Dict{Int64,Array{Any}}()

#modelvars = JuMP.read_from_file("../../FinalProject/storage_expansion_revised/second_stage/noint_PR_exp5p0015_scen_1.mps")
modelvars = JuMP.read_from_file("./storage_expansion_revised/ercot/PR_exp3_scen_1.mps")

firstvars = []
for bus in buses
    push!(firstvars,(string(get_PR_variable(modelvars,bus)), 0.0, Inf, 0.0))
    #push!(firstvars,(string(get_ER_variable(modelvars,bus)), 0.0, Inf, 0.0))
end

vardict[1] = firstvars

secondvars = [];

for ts in timesteps
    
    for gen in gens
        push!(secondvars,string(get_thermal_variable(modelvars,gen,ts)))
    end
    for br in branches
        push!(secondvars,string(get_line_variable(modelvars,br,ts)))
    end
    
    push!(secondvars, string(get_wind_variable(modelvars,122,ts)))
    
    for bus in buses
        push!(secondvars, string(get_charging_variable(modelvars,bus,ts)))
        push!(secondvars, string(get_discharging_variable(modelvars,bus,ts)))
        push!(secondvars, string(get_stored_variable(modelvars,bus,ts)))
        push!(secondvars, string(get_lossofload_variable(modelvars,bus,ts)))
        push!(secondvars, string(get_overload_variable(modelvars,bus,ts)))
    end
end

vardict[2] = secondvars

#=
firststagevars = Dict()

for index in 1:length(vardict[1])
    var = vardict[1][index]
    firststagevars[var[1]] = LShaped.FirstStageVariableInfo(var[1], index, var[4], nothing, var[2], var[3])
end

tol = 1e-6
#empty dict temporary until I can figure out how to store things.
subprob = Dict();

firststage = LShaped.FirstStageInfo(firststagevars, subprob, dataloc);=#


model = JuMP.read_from_file("./storage_expansion_revised/first_stage/first_stage_model_PR_exp1B.mps")
JuMP.set_optimizer(model, Gurobi.Optimizer)
set_optimizer_attribute(model, "NumericFocus", 3)
#JuMP.set_optimizer_attribute(model, "Method", 2)
set_optimizer_attribute(model, "OutputFlag", 1)
add_theta_to_objective!(model)

scen = 6803

#println("Adjusting to scenario $(ercotscens[i])")
wind = (1/wmax)*(wrts/100)*collect(winddf[scen,3:26])
load = (1/lmax)*(1.35*lrts/100)*collect(loaddf[scen,3:26])

for bus in buses
    lf = loaddis[bus]
    for ts in timesteps
        con = get_load_balance(model2nd, bus, ts)
        oldval = JuMP.constraint_object(con).set.value
        lval = load[ts]
        JuMP.set_normalized_rhs(con, lf*lval)
        newval = JuMP.constraint_object(con).set.value
        #println("$(name(con)), $(oldval), $(newval)")
    end
end

# change ptdf constraint (remember to run load changes FIRST)
for ts in timesteps
    for br in branches
        ptdfcon = get_ptdf_con(model2nd,br,ts)

        valold = JuMP.constraint_object(ptdfcon).set.value
        valnew = 0.0
        for bus in buses
            buscon = get_load_balance(model2nd,bus,ts)

            loadcon = copy(JuMP.constraint_object(buscon).func)
            loadval = copy(JuMP.constraint_object(buscon).set.value)

            valnew -= ptdfdict[br][bus]*loadval

        end 

        JuMP.set_normalized_rhs(ptdfcon, valnew)
        #println("$(JuMP.name(ptdfcon)), $(valold), $(valnew)")

    end
end

bus = 122
for ts in timesteps
    con = get_wind_ub(model2nd, bus, ts)
    oldval = JuMP.constraint_object(con).set.upper
    wval = wind[ts]
    JuMP.set_normalized_rhs(con, wval)
    newval = JuMP.constraint_object(con).set.upper
    #println("$(name(con)), $(oldval), $(newval)")
end

add_wc_variables_to_fs_model!(model, vardict, scen)
v1array = get_1st_stage_variable_array(vardict)
add_wc_constraints!(model, model2nd, scen, v1array)
add_obj_constraint!(model, model2nd, scen, v1array)

optimize!(model)
objective_value(model)

model

JuMP.set_optimizer_attribute(model, "Method", 1)


for i = 101:124
    println(value(get_PR_variable(model,i)))
end

println(value(variable_by_name(model,"theta")))

get_load_balance(model, 101,19, scen)

scen
