==> jn setups/Step 21_1 get closest to center.ipynb <==
using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames
using Random
using LinearAlgebra
using Clustering

include("../LShaped.jl/src/structs.jl")
include("../LShaped.jl/src/setup.jl")
include("../LShaped.jl/src/algorithm.jl")

include("./parameters.jl")
include("./get_functions.jl")
include("./modification_functions_async.jl")

loadcsv = CSV.File("./LOAD.csv");
ptdfdf = DataFrame(CSV.File("./ptdfsmall.csv"));
loaddf = DataFrame(CSV.File("./loaddata.csv"))
winddf = DataFrame(CSV.File("./winddata.csv"))

part1458 = DataFrame(CSV.File("./scenarios/part1458.csv"))

let 
    lmax = 0.0
    wmax = 0.0
    for i = 1:24
        #println(i, " ", maximum(loaddf[:,2+i]))
        if maximum(loaddf[:,2+i]) > lmax
            lmax = maximum(loaddf[:,2+i])
        end
        #println(maximum(winddf[:,2+i]))
        if maximum(winddf[:,2+i]) > wmax
            wmax = maximum(winddf[:,2+i])
        end
    end
    global lmax = lmax
    global wmax = wmax
end

loaddis = load_distribution_dict(loadcsv);

ptdfdict = Dict()

for i = 1:38
    br = ptdfdf[i,1]
    ptdfdict[br] = Dict()
    for j = 2:25
        bus = parse(Int64,names(ptdfdf)[j])
        ptdfdict[br][bus] = ptdfdf[i,j]
    end
end

lrts = 2850.0
wrts = 713.5

loadmat = lrts/lmax*1.35*transpose(Matrix(loaddf[:,3:26]))
windmat = wrts/wmax*transpose(Matrix(winddf[:,3:26]))

X = loadmat - windmat
X = 1/maximum(X)*X

K = 12
exp = 1

trvec = collect(part1458[exp,:])

X1458 = Array{Float64}(undef, size(X,1), 1458)
for i = 1:1458
    X1458[:,i] = X[:, trvec[i]]
end
X1458

centers = DataFrame(CSV.File("./scenarios/centers_$(K)_$(exp).csv"))

cldict = CSV.File("./scenarios/cldict_$(K)_$(exp).csv") |> Dict

clid = 1
center = centers[:,clid]

vect = Int64[];
n = length(cldict[clid])
string = cldict[clid][2:n-1]
for i = 1:length(string)
    if string[i] == ','
        if i-1 == 1
            push!(vect, parse(Int64, string[i-1]))
        elseif i-2 == 1
            push!(vect, parse(Int64, string[i-2:i-1]))
        elseif i-3 == 1
            push!(vect, parse(Int64, string[i-3:i-1]))
        elseif i-4 == 1
            push!(vect, parse(Int64, string[i-4:i-1]))
        elseif string[i-2] == ' ' 
            push!(vect,parse(Int64, string[i-1]))
        elseif string[i-3] == ' '
            push!(vect,parse(Int64, string[i-2:i-1]))
        elseif string[i-4] == ' '
            push!(vect,parse(Int64, string[i-3:i-1]))
        elseif string[i-5] == ' '
            push!(vect,parse(Int64, string[i-4:i-1]))
        else
            println("Uh oh!")
        end
    elseif i == n-2
        if i == 1
            push!(vect, parse(Int64, string[i]))
        elseif i-1 == 1
            push!(vect, parse(Int64, string[i-1:i]))
        elseif i-2 == 1
            push!(vect, parse(Int64, string[i-2:i]))
        elseif i-3 == 1
            push!(vect, parse(Int64, string[i-3:i]))
        elseif string[i-1] == ' ' 
            push!(vect,parse(Int64, string[i]))
        elseif string[i-2] == ' '
            push!(vect,parse(Int64, string[i-1:i]))
        elseif string[i-3] == ' '
            push!(vect,parse(Int64, string[i-2:i]))
        elseif string[i-4] == ' '
            push!(vect,parse(Int64, string[i-3:i]))
        else
            println("Uh oh!")
        end
            
    end
end


distvect = Vector{Float64}(undef, length(vect))
for i = 1:length(vect)
    dist = norm(center-X[:,vect[i]])
    println(i, " ",dist)
    distvect[i] = dist
end

#vect[argmin(distvect)]
println(K, " ", exp, " ", clid, " ", vect[argmin(distvect)])

X[:,vect[25]]

minimum(distvect)

for K in [12,24]
    for exp = 1:5

        trvec = collect(part1458[exp,:])

        X1458 = Array{Float64}(undef, size(X,1), 1458)
        for i = 1:1458
            X1458[:,i] = X[:, trvec[i]]
        end

        centers = DataFrame(CSV.File("./scenarios/centers_$(K)_$(exp).csv"))
        cldict = CSV.File("./scenarios/cldict_$(K)_$(exp).csv") |> Dict

        for clid = 1:K

            center = centers[:,clid]

            vect = Int64[];
            n = length(cldict[clid])
            string = cldict[clid][2:n-1]
            for i = 1:length(string)
                if string[i] == ','
                    if i-1 == 1
                        push!(vect, parse(Int64, string[i-1]))
                    elseif i-2 == 1
                        push!(vect, parse(Int64, string[i-2:i-1]))
                    elseif i-3 == 1
                        push!(vect, parse(Int64, string[i-3:i-1]))
                    elseif i-4 == 1
                        push!(vect, parse(Int64, string[i-4:i-1]))
                    elseif string[i-2] == ' ' 
                        push!(vect,parse(Int64, string[i-1]))
                    elseif string[i-3] == ' '
                        push!(vect,parse(Int64, string[i-2:i-1]))
                    elseif string[i-4] == ' '
                        push!(vect,parse(Int64, string[i-3:i-1]))
                    elseif string[i-5] == ' '
                        push!(vect,parse(Int64, string[i-4:i-1]))
                    else
                        println("Uh oh!")
                    end
                elseif i == n-2
                    if i == 1
                        push!(vect, parse(Int64, string[i]))
                    elseif i-1 == 1
                        push!(vect, parse(Int64, string[i-1:i]))
                    elseif i-2 == 1
                        push!(vect, parse(Int64, string[i-2:i]))
                    elseif i-3 == 1
                        push!(vect, parse(Int64, string[i-3:i]))
                    elseif string[i-1] == ' ' 
                        push!(vect,parse(Int64, string[i]))
                    elseif string[i-2] == ' '
                        push!(vect,parse(Int64, string[i-1:i]))
                    elseif string[i-3] == ' '
                        push!(vect,parse(Int64, string[i-2:i]))
                    elseif string[i-4] == ' '
                        push!(vect,parse(Int64, string[i-3:i]))
                    else
                        println("Uh oh!")
                    end

                end
            end

            distvect = Vector{Float64}(undef, length(vect))
            for i = 1:length(vect)
                dist = norm(center-X[:,vect[i]])
                #println(i, " ",dist)
                distvect[i] = dist
            end

            #vect[argmin(distvect)]
            println(K, ", ", exp, ", ", clid, ", ", vect[argmin(distvect)], ", ", minimum(distvect))
        end
    end
end
