==> jn setups/Step 22 construct four_wind model.ipynb <==
using JuMP
using DataFrames
using CSV
using DataFrames
using Random
#using LShaped
include("../../LShaped.jl/src/structs.jl")
include("../../LShaped.jl/src/setup.jl")
include("../../LShaped.jl/src/algorithm.jl")

## make sure these files and their dependencies are in their proper place ##
include("../parameters.jl")
include("../get_functions.jl")
include("../modification_functions.jl")
##

## upload these to cluster ##
loaddf = DataFrame(CSV.File("../four_wind_turbine_experiments/loaddata.csv"))
wind103df = DataFrame(CSV.File("../four_wind_turbine_experiments/winddata103.csv"))
wind109df = DataFrame(CSV.File("../four_wind_turbine_experiments/winddata109.csv"))
wind117df = DataFrame(CSV.File("../four_wind_turbine_experiments/winddata117.csv"))
wind122df = DataFrame(CSV.File("../four_wind_turbine_experiments/winddata122.csv"))
##

winddfdict = Dict(103 => wind103df, 109 => wind109df, 117 => wind117df, 122 => wind122df)

oldmodel = JuMP.read_from_file("../four_wind_turbine_experiments/four_wind/old_second_stage_model_scen_1.mps")

for wbus in wind_buses
    wind = (1/wmaxdict[wbus])*(wrtsdict[wbus]/100)*collect(winddfdict[wbus][1,3:26])
    if wbus != 122
        for ts in timesteps
            var = @variable(oldmodel, base_name = string("Pre_{", wbus, "_WIND_1,",ts,"}"))
            @constraint(oldmodel, var >= 0.0, base_name = string("Wlb_{",wbus,",",ts,"}"))
            @constraint(oldmodel, var <= wind[ts], base_name = string("Wub_{",wbus,",",ts,"}"))
        end
    end
end

get_wind_ub(oldmodel, 103, 9)

# to do:
# - attach it to load balance (only at relevent buses at each time step)
# - attach it to ptdf constraint (this is already sorted by bus iirc, no need for hacky fixes)
# thus this will extend into 3/15 at least.

for ts in timesteps 
    con = get_load_balance(oldmodel, 103, ts)
    var = get_wind_variable(oldmodel, 103, ts)
    set_normalized_coefficient(con, var, 1.0)
end

for ts in timesteps 
    con = get_load_balance(oldmodel, 109, ts)
    var = get_wind_variable(oldmodel, 109, ts)
    set_normalized_coefficient(con, var, 1.0)
end

for ts in timesteps 
    con = get_load_balance(oldmodel, 117, ts)
    var = get_wind_variable(oldmodel, 117, ts)
    set_normalized_coefficient(con, var, 1.0)
end

ptdfdf[, Symbol(102)]

ptdfdf

branchtorow = Dict()
count = 0
for br in branches
    for i = 1:38
        if ptdfdf[i,1] == br
            branchtorow[br] = i
            count += 1
            println(count)
        end
    end
end

for br in branches
    coeff103 = -ptdfdf[branchtorow[br], Symbol(103)]
    coeff109 = -ptdfdf[branchtorow[br], Symbol(109)]
    coeff117 = -ptdfdf[branchtorow[br], Symbol(117)]
    for ts in timesteps
        con = get_ptdf_con(oldmodel, br, ts)
        var103 = get_wind_variable(oldmodel, 103, ts)
        var109 = get_wind_variable(oldmodel, 109, ts)
        var117 = get_wind_variable(oldmodel, 117, ts)
        
        set_normalized_coefficient(con, var103, coeff103)
        set_normalized_coefficient(con, var109, coeff109)
        set_normalized_coefficient(con, var117, coeff117)
    end
end


write_to_file(oldmodel, "../four_wind_turbine_experiments/four_wind/second_stage_model_scen_1.mps")

omodel = JuMP.read_from_file("../four_wind_turbine_experiments/four_wind/old_second_stage_model_scen_1.mps")
nmodel = JuMP.read_from_file("../four_wind_turbine_experiments/four_wind/second_stage_model_scen_1.mps")

omodel

using Gurobi

JuMP.set_optimizer(omodel, Gurobi.Optimizer)
JuMP.set_optimizer(nmodel, Gurobi.Optimizer)

optimize!(omodel)

optimize!(nmodel)
