==> jn setups/Step 24 Preprocess data for clusters and add clusters maybe 4wind.ipynb <==
using JuMP
using DataFrames
using CSV
using Gurobi
using DataFrames
using Random
using LinearAlgebra
using Clustering

include("../../LShaped.jl/src/structs.jl")
include("../../LShaped.jl/src/setup.jl")
include("../../LShaped.jl/src/algorithm.jl")

include("../parameters.jl")
include("../get_functions.jl")
include("../modification_functions_async.jl")

loadcsv = CSV.File("../RTS_data_summary/LOAD.csv");
ptdfdf = DataFrame(CSV.File("../RTS_data_summary/ptdfsmall.csv"));
loaddf = DataFrame(CSV.File("../four_wind_turbine_experiments/loaddata.csv"))
winddfdict = Dict()
for wb in wind_buses
    winddfdict[wb] = DataFrame(CSV.File("../four_wind_turbine_experiments/winddata$(wb).csv"))
end
part1458 = DataFrame(CSV.File("../scenarios/part1458.csv"))

loaddis = load_distribution_dict(loadcsv);

ptdfdict = Dict()

for i = 1:38
    br = ptdfdf[i,1]
    ptdfdict[br] = Dict()
    for j = 2:25
        bus = parse(Int64,names(ptdfdf)[j])
        ptdfdict[br][bus] = ptdfdf[i,j]
    end
end


loadmat = lrts/lmax*1.35*transpose(Matrix(loaddf[:,3:26]))

windmat = zeros(24,6940)

for wb in wind_buses
    windmat += wrtsdict[wb]/wmaxdict[wb]*transpose(Matrix(winddfdict[wb][:,3:26]))
end

X = loadmat - windmat

X = X - minimum(X)*ones(24,6940)

X = 1/maximum(X)*X

maximum(X)

X1458_1 = Array{Float64}(undef, size(X,1), 1458)

trvec = collect(part1458[1,:])

for i = 1:1458
    X1458_1[:,i] = X[:, trvec[i]]
end
X1458_1

Random.seed!(7)
C = Clustering.kmeans(X1458_1,12)

fieldnames(typeof(C))

C.assignments[1] #10
C.centers[:,10]
norm(X1458_1[:,1]-C.centers[:,10])^2- C.costs[1]

fieldnames(typeof(C))

cldict = Dict()
for i = 1:12
    cldict[i] = Int64[]
end
    

zeros(Int64,5)

lowest_cost = Inf*ones(12)
closest_center = zeros(Int64,12)

for i = 1:1458
    if C.costs[i] < lowest_cost[C.assignments[i]]
        closest_center[C.assignments[i]] = trvec[i]
        lowest_cost[C.assignments[i]] = C.costs[i]
    end
end
println(closest_center)
        

for i = 1:1458
    push!(cldict[C.assignments[i]], trvec[i])
end

lowest = Inf
best = 0
for i = 1:81
    println(cldict[12][i], " ", norm(X[:,cldict[12][i]]-C.centers[:,12]))
    if  norm(X[:,cldict[12][i]]-C.centers[:,12]) < lowest
        lowest = norm(X[:,cldict[12][i]]-C.centers[:,12])
        best = cldict[12][i]
    end
end

best

CSV.write("./scenarios/cldict_24_2.csv",cldict)

show(C.counts)

CSV.write("./scenarios/centers_24_2.csv", DataFrame(C.centers))

for test in [420, 69]
    println(test)
    show(C.counts)
end
typeof(C.centers)

X1458 = Array{Float64}(undef, size(X,1), 1458)

for exp = 1:5
    
    trvec = collect(part1458[exp,:])
    
    for i = 1:1458
        X1458[:,i] = X[:, trvec[i]]
    end

    for K in [12,24]
        Random.seed!(7)
        C = Clustering.kmeans(X1458,K)
        
        cldict = Dict()
        for i = 1:K
            cldict[i] = Int64[]
        end
        for i = 1:1458
            push!(cldict[C.assignments[i]], trvec[i])
        end
        
        lowest_cost = Inf*ones(K)
        closest_center = zeros(Int64,1,K)

        for i = 1:1458
            if C.costs[i] < lowest_cost[C.assignments[i]]
                closest_center[1,C.assignments[i]] = trvec[i]
                lowest_cost[C.assignments[i]] = C.costs[i]
            end
        end
        
        CSV.write("../scenarios/four_wind/cldict_$(K)_$(exp).csv",cldict)
        CSV.write("../scenarios/four_wind/centers_$(K)_$(exp).csv", DataFrame(C.centers))
        CSV.write("../scenarios/four_wind/closest_to_center_$(K)_$(exp).csv", DataFrame(closest_center))
        
        println(K, " ", exp)
        show(C.counts)
        println(closest_center)
        println()
    end
end
        
        
        
    

zeros(Int64,1,2)
